EAA Adjust for AuditLog by sendmail
