+++ Backend/src/controllers/workOrder.controller.js
@@ -57,7 +57,7 @@ const createWorkOrder = async (req, res) => {
                 html: `<p>Por favor gestionar la orden de trabajo asignada ${newWorkOrder.workOrderDescription}; Tipo de servicio: ${newWorkOrder.serviceType}; Fecha programada: ${newWorkOrder.workOrderScheduledDate}; Duración estimada: ${newWorkOrder.workOrderEstimatedDuration} horas; Dirección de la orden de trabajo: ${newWorkOrder.workOrderAddress}; Geolocalizacion: ${newWorkOrder.workOrderLocation}</p><br><p>Datos del cliente: email ${newWorkOrder.workOrderclientEmail}; persona de contacto ${newWorkOrder.workOrderClientContactPerson}; teléfono de contacto ${newWorkOrder.workOrderClientPhone}</p>`
             }
             // Reutilizar la función de envío de correos
-            const result = await mail.sendEmail(emailData);
+            const result = await mail.sendEmail(req, 'createWorkOrder', newWorkOrder._id, emailData);
             console.log('result sendMail', result);
             if (!result.success) {
                 return res.status(201).json({ ok: true, message: 'Orden de trabajo creada exitosamente. No fue posible enviar correo al tecnico asignado.', data: newWorkOrder });
